pyproject.toml
[project]
name = "create-nsfw-image"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "opencv-python>=4.11.0.86",
    "transformers>=4.49.0",
    "pillow>=11.1.0",
]
